cleanup in Layer class
